refactor: refactor all dependencies
