+++ src/chatting/chatting.module.ts
@@ -1,25 +1,22 @@
 import { AgoraModule } from '../agora/agora.module';
 import { dynamooseModule } from '../config.dynamoose';
-import { QuestionRepository } from '../question/question.repository';
-import { TutoringRepository } from '../tutoring/tutoring.repository';
-import { UploadService } from '../upload/upload.service';
-import { UserRepository } from '../user/user.repository';
+import { SocketModule } from '../socket/socket.module';
+import { UserModule } from '../user/user.module';
 import { ChattingController } from './chatting.controller';
 import { ChattingRepository } from './chatting.repository';
 import { ChattingService } from './chatting.service';
 import { Module } from '@nestjs/common';
 
 @Module({
-  imports: [dynamooseModule, ChattingModule, AgoraModule],
-  controllers: [ChattingController],
-  providers: [
-    ChattingService,
-    ChattingRepository,
-    UserRepository,
-    QuestionRepository,
-    UploadService,
-    TutoringRepository,
+  imports: [
+    dynamooseModule,
+    ChattingModule,
+    AgoraModule,
+    UserModule,
+    SocketModule,
   ],
+  controllers: [ChattingController],
+  providers: [ChattingService, ChattingRepository],
   exports: [ChattingRepository],
 })
 export class ChattingModule {}
