+++ src/event/event.module.ts
@@ -1,13 +1,13 @@
 import { dynamooseModule } from '../config.dynamoose';
-import { UploadService } from '../upload/upload.service';
+import { UploadModule } from '../upload/upload.module';
 import { EventController } from './event.controller';
 import { EventRepository } from './event.repository';
 import { EventService } from './event.service';
 import { Module } from '@nestjs/common';
 
 @Module({
-  imports: [dynamooseModule],
+  imports: [dynamooseModule, UploadModule],
   controllers: [EventController],
-  providers: [EventService, EventRepository, UploadService],
+  providers: [EventService, EventRepository],
 })
 export class EventModule {}
