+++ src/follow/follow.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { FollowController } from './follow.controller';
+
+describe('FollowController', () => {
+  let controller: FollowController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [FollowController],
+    }).compile();
+
+    controller = module.get<FollowController>(FollowController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
