+++ src/follow/follow.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { FollowService } from './follow.service';
+
+describe('FollowService', () => {
+  let service: FollowService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [FollowService],
+    }).compile();
+
+    service = module.get<FollowService>(FollowService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
