+++ src/tutoring/tutoring.controller.ts
@@ -1,4 +1,10 @@
 import { AccessToken } from '../auth/entities/auth.entity';
+import {
+  TeacherOperation,
+  UserOperation,
+} from '../user/descriptions/user.operation';
+import { UserParam } from '../user/descriptions/user.param';
+import { UserResponse } from '../user/descriptions/user.response';
 import { TutoringOperation } from './descriptions/tutoring.operation';
 import { TutoringResponse } from './descriptions/tutoring.response';
 import { CreateReviewDto } from './dto/create-review.dto';
@@ -8,6 +14,7 @@ import { Body, Controller, Get, Headers, Param, Post } from '@nestjs/common';
 import {
   ApiBearerAuth,
   ApiOperation,
+  ApiParam,
   ApiResponse,
   ApiTags,
 } from '@nestjs/swagger';
@@ -100,5 +107,26 @@ export class TutoringController {
       tutoringId,
       createReviewDto,
     );
+  }
+
+  @ApiTags('User')
+  @ApiBearerAuth('Authorization')
+  @ApiOperation(UserOperation.tutoringList)
+  @Get('user/tutoring/list')
+  @ApiResponse(UserResponse.tutoringList)
+  tutoringList(@Headers() headers: Headers) {
+    return this.tutoringService.tutoringList(AccessToken.userId(headers));
+  }
+
+  @ApiTags('Teacher')
+  @ApiBearerAuth('Authorization')
+  @ApiParam(UserParam.teacherId)
+  @ApiOperation(TeacherOperation.reviewList)
+  @Get('teacher/review/list/:teacherId')
+  reviewList(
+    @Headers() headers: Headers,
+    @Param('teacherId') teacherId: string,
+  ) {
+    return this.tutoringService.reviewList(teacherId);
   }
 }
