+++ src/user/user.controller.ts
@@ -1,10 +1,9 @@
 import { AccessToken } from '../auth/entities/auth.entity';
-import { TeacherOperation, UserOperation } from './descriptions/user.operation';
+import { UserOperation } from './descriptions/user.operation';
 import { UserParam } from './descriptions/user.param';
 import { UserResponse } from './descriptions/user.response';
 import { CreateStudentDto, CreateTeacherDto } from './dto/create-user.dto';
 import { LoginUserDto } from './dto/login-user.dto';
-import { SetFCMTokenUserDto } from './dto/setFCMToken-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserService } from './user.service';
 import { Body, Controller, Get, Headers, Param, Post } from '@nestjs/common';
@@ -75,26 +74,6 @@ export class UserController {
 
   @ApiTags('User')
   @ApiBearerAuth('Authorization')
-  @ApiParam(UserParam.userId)
-  @ApiOperation(UserOperation.otherFollowers)
-  @ApiResponse(UserResponse.followInfo)
-  @Get('user/followers/:userId')
-  otherFollowers(@Headers() headers: Headers, @Param('userId') userId: string) {
-    return this.userService.otherFollowers(userId);
-  }
-
-  @ApiTags('User')
-  @ApiBearerAuth('Authorization')
-  @ApiParam(UserParam.userId)
-  @ApiOperation(UserOperation.otherFollowing)
-  @ApiResponse(UserResponse.followInfo)
-  @Get('user/following/:userId')
-  otherFollowing(@Headers() headers: Headers, @Param('userId') userId: string) {
-    return this.userService.otherFollowing(userId);
-  }
-
-  @ApiTags('User')
-  @ApiBearerAuth('Authorization')
   @ApiOperation(UserOperation.me.withdraw)
   @ApiResponse(UserResponse.me.withdraw)
   @Get('user/withdraw')
@@ -103,40 +82,6 @@ export class UserController {
       AccessToken.userId(headers),
       AccessToken.authorization(headers),
     );
-  }
-
-  @ApiTags('Student')
-  @ApiBearerAuth('Authorization')
-  @ApiParam(UserParam.userId)
-  @ApiOperation(UserOperation.follow)
-  @Get('student/follow/:userId')
-  follow(@Headers() headers: Headers, @Param('userId') userId: string) {
-    return this.userService.follow(AccessToken.userId(headers), userId);
-  }
-
-  @ApiTags('Student')
-  @ApiBearerAuth('Authorization')
-  @ApiParam(UserParam.userId)
-  @ApiOperation(UserOperation.unfollow)
-  @Get('student/unfollow/:userId')
-  unfollow(@Headers() headers: Headers, @Param('userId') userId: string) {
-    return this.userService.unfollow(AccessToken.userId(headers), userId);
-  }
-
-  @ApiTags('Student')
-  @ApiBearerAuth('Authorization')
-  @ApiOperation(UserOperation.following)
-  @Get('student/following')
-  following(@Headers() headers: Headers) {
-    return this.userService.following(AccessToken.userId(headers));
-  }
-
-  @ApiTags('Teacher')
-  @ApiBearerAuth('Authorization')
-  @ApiOperation(UserOperation.followers)
-  @Get('teacher/followers')
-  followers(@Headers() headers: Headers) {
-    return this.userService.followers(AccessToken.userId(headers));
   }
 
   @ApiTags('User')
@@ -149,51 +94,8 @@ export class UserController {
 
   @ApiTags('User')
   @ApiBearerAuth('Authorization')
-  @ApiOperation(UserOperation.onlineTeacher)
-  @ApiResponse(UserResponse.onlineTeacher)
-  @Get('user/list/teacher/online')
-  getOnlineTeachers(@Headers() headers: Headers) {
-    return this.userService.getOnlineTeachers(AccessToken.userId(headers));
-  }
-
-  @ApiTags('User')
-  @ApiBearerAuth('Authorization')
-  @Post('user/fcmToken')
-  setFCMToken(
-    @Headers() headers: Headers,
-    @Body() setFCMTokenUserDto: SetFCMTokenUserDto,
-  ) {
-    return this.userService.setFCMToken(
-      AccessToken.userId(headers),
-      setFCMTokenUserDto.fcmToken,
-    );
-  }
-
-  @ApiTags('User')
-  @ApiBearerAuth('Authorization')
   @Get('user/receiveFreeCoin')
   receiveFreeCoin(@Headers() headers: Headers) {
     return this.userService.receiveFreeCoin(AccessToken.userId(headers));
-  }
-
-  @ApiTags('User')
-  @ApiBearerAuth('Authorization')
-  @ApiOperation(UserOperation.tutoringList)
-  @Get('user/tutoring/list')
-  @ApiResponse(UserResponse.tutoringList)
-  tutoringList(@Headers() headers: Headers) {
-    return this.userService.tutoringList(AccessToken.userId(headers));
-  }
-
-  @ApiTags('Teacher')
-  @ApiBearerAuth('Authorization')
-  @ApiParam(UserParam.teacherId)
-  @ApiOperation(TeacherOperation.reviewList)
-  @Get('teacher/review/list/:teacherId')
-  reviewList(
-    @Headers() headers: Headers,
-    @Param('teacherId') teacherId: string,
-  ) {
-    return this.userService.reviewList(teacherId);
   }
 }
