+++ src/user/user.module.ts
@@ -1,25 +1,16 @@
-import { AgoraModule } from '../agora/agora.module';
 import { AuthModule } from '../auth/auth.module';
 import { dynamooseModule } from '../config.dynamoose';
-import { QuestionRepository } from '../question/question.repository';
 import { RedisModule } from '../redis/redis.module';
-import { TutoringRepository } from '../tutoring/tutoring.repository';
-import { UploadService } from '../upload/upload.service';
+import { UploadModule } from '../upload/upload.module';
 import { UserController } from './user.controller';
 import { UserRepository } from './user.repository';
 import { UserService } from './user.service';
 import { Module } from '@nestjs/common';
 
 @Module({
-  imports: [dynamooseModule, AuthModule, RedisModule, AgoraModule],
+  imports: [dynamooseModule, AuthModule, RedisModule, UploadModule],
   controllers: [UserController],
-  providers: [
-    UserService,
-    UserRepository,
-    UploadService,
-    TutoringRepository,
-    QuestionRepository,
-  ],
-  exports: [UserRepository],
+  providers: [UserService, UserRepository],
+  exports: [UserRepository, UserService],
 })
 export class UserModule {}
